+++ src/main/java/defs/objects/flyer/EnemyFlyer.java
@@ -7,27 +7,27 @@ import main.Main;
 import main.StatsCollector;
 import temperature.Functions;
 
-public class EnemyFlyer extends Flyer{
+public class EnemyFlyer extends Flyer {
 
-	private int deltaX,acceleration;
+	private int deltaX, acceleration;
 	private int[] force;
 	private double intelligence;
 	final double AbsVel;
-	
+
 	public EnemyFlyer(Main main) {
-		super(main, (int)(main.Width/2.0-((new Random()).nextInt())%(main.Width/6)), 200);
+		super(main, (int) (main.Width / 2.0 - ((new Random()).nextInt()) % (main.Width / 6)), 200);
 		setMass(300);
-		setIntelligence(1.0 + (new Random()).nextInt(50)/100.0);
-		setSpeedY(((int)((5+(new Random()).nextInt()%5)*main.Height/1080)));
-		AbsVel=getIntelligence();
-		this.setSize(60*main.Height/1080);
+		setIntelligence(1.0 + (new Random()).nextInt(50) / 100.0);
+		setSpeedY(((5 + (new Random()).nextInt() % 5) * main.Height / 1080));
+		AbsVel = getIntelligence();
+		this.setSize(60 * main.Height / 1080);
 	}
 
-	@Override 
+	@Override
 	public void move() {
-	      setPosition(Functions.mathOperator.AdditionOfVectors(getPosition(), getVelocity()));
-	    } 
-	
+		setPosition(Functions.mathOperator.AdditionOfVectors(getPosition(), getVelocity()));
+	}
+
 	@Override
 	public void move(Main main) {
 		if ((new Random()).nextInt(1000) > 950) {
@@ -40,20 +40,19 @@ public class EnemyFlyer extends Flyer{
 //		}
 //		this.setVelocity(Functions.ScalarMultiplication(intelligence*0.01/AbsVel,
 //				Functions.AdditionOfVectors(Functions.ReversalOfVector(this.getPosition()),((HanSolo) (main.getHanSolo())).getPosition())));
-		setX(getX()+(int)(Math.pow(getIntelligence(),1)*getSpeedX()));
-		setAcceleration(main.getHanSolo().getX()-getX());
-		double[]vec1 = new double[2],vec2=new double[2];
-		vec1[0]=main.getHanSolo().getX();
-		vec1[1]=main.Height/8;
-		vec2[0]=this.getX();
-		vec2[1]=this.getY();
-		double[] frc=Functions.mathOperator.AdditionOfVectors(vec1,
-				Functions.mathOperator.ReversalOfVector(vec2));
-		double tmpLength=Functions.mathOperator.MagnitudeOfVector(frc);
-		if (tmpLength>0.1){
-			frc=Functions.mathOperator.ScalarMultiplication(0.1,Functions.mathOperator.UnitVector(frc));
+		setX(getX() + (int) (Math.pow(getIntelligence(), 1) * getSpeedX()));
+		setAcceleration(main.getHanSolo().getX() - getX());
+		double[] vec1 = new double[2], vec2 = new double[2];
+		vec1[0] = main.getHanSolo().getX();
+		vec1[1] = main.Height / 8;
+		vec2[0] = this.getX();
+		vec2[1] = this.getY();
+		double[] frc = Functions.mathOperator.AdditionOfVectors(vec1, Functions.mathOperator.ReversalOfVector(vec2));
+		double tmpLength = Functions.mathOperator.MagnitudeOfVector(frc);
+		if (tmpLength > 0.1) {
+			frc = Functions.mathOperator.ScalarMultiplication(0.1, Functions.mathOperator.UnitVector(frc));
 		}
-		this.setVelocity(Functions.mathOperator.AdditionOfVectors(frc,this.getVelocity()));
+		this.setVelocity(Functions.mathOperator.AdditionOfVectors(frc, this.getVelocity()));
 //		if(Math.abs(this.getPosition()[0]-((HanSolo)(main.getHanSolo())).getPosition()[0])>50) 
 //		{
 //			if (getAcceleration()>0) {
@@ -66,27 +65,25 @@ public class EnemyFlyer extends Flyer{
 //		getVelocity()[1]=Math.max((int)(3*intelligence),3);
 		move();
 		// Do we collide?
-		if (this.getY()+this.getSize()>main.getHanSolo().getY() && 
-			this.getX()<main.getHanSolo().getX()+main.getHanSolo().getSize() &&
-			this.getX()+this.getSize()>main.getHanSolo().getX() &&
-			this.getY()+this.getSize()<main.getHanSolo().getY()+main.getHanSolo().getSize() &&
-			this.getY()+this.getSize()<main.Height){
-			((HanSolo)main.getHanSolo()).addScore(50);
+		if (this.getY() + this.getSize() > main.getHanSolo().getY()
+				&& this.getX() < main.getHanSolo().getX() + main.getHanSolo().getSize()
+				&& this.getX() + this.getSize() > main.getHanSolo().getX()
+				&& this.getY() + this.getSize() < main.getHanSolo().getY() + main.getHanSolo().getSize()
+				&& this.getY() + this.getSize() < main.Height) {
+			((HanSolo) main.getHanSolo()).addScore(50);
 			selfDestroy(main);
-			StatsCollector.getInstance().setEnemies(
-					1+StatsCollector.getInstance().getEnemies()
-					);
-			Explosion expl=new Explosion(main, this.getX(), this.getY());
-			expl.setSize(this.getSize()*2);
+			StatsCollector.getInstance().setEnemies(1 + StatsCollector.getInstance().getEnemies());
+			Explosion expl = new Explosion(main, this.getX(), this.getY());
+			expl.setSize(this.getSize() * 2);
 			main.add(expl);
-			((HanSolo)main.getHanSolo()).setHealth(((HanSolo)main.getHanSolo()).getHealth()-30);
+			((HanSolo) main.getHanSolo()).setHealth(((HanSolo) main.getHanSolo()).getHealth() - 30);
 		}
-		if (this.getY()>main.Height){
-			((HanSolo)main.getHanSolo()).addScore(-25);
+		if (this.getY() > main.Height) {
+			((HanSolo) main.getHanSolo()).addScore(-25);
 			selfDestroy(main);
 		}
 	}
-	
+
 	public void setIntelligence(double intelligence) {
 		this.intelligence = intelligence;
 	}
@@ -98,7 +95,7 @@ public class EnemyFlyer extends Flyer{
 	public void setDeltaX(int deltaX) {
 		this.deltaX = deltaX;
 	}
-	
+
 	@Override
 	public void shoot(Main main) {
 		this.getGun().shoot(main);
