+++ src/main/java/defs/objects/guns/DoublePlainGun.java
@@ -4,8 +4,8 @@ import defs.objects.flyer.Flyer;
 import defs.objects.munition.Bullet;
 import main.Main;
 
-public class DoublePlainGun extends PlainGun{
-	
+public class DoublePlainGun extends PlainGun {
+
 	public DoublePlainGun(Flyer fly) {
 		super(fly);
 		setVerbrauch(6);
@@ -13,10 +13,12 @@ public class DoublePlainGun extends PlainGun{
 
 	@Override
 	public void shoot(Main main) {
-		main.add(new Bullet(main,this.getFlyer().getX()-(int)(0.5*getFlyer().getSize())+(int)(this.getFlyer().getSize()/3),
-				this.getFlyer().getY()-20,(int)this.getFlyer().getSpeedX(),20));
-		main.add(new Bullet(main,this.getFlyer().getX()-(int)(0.5*getFlyer().getSize())+2*(int)(this.getFlyer().getSize()/3),
-				this.getFlyer().getY()-20,(int)this.getFlyer().getSpeedX(),20));
+		main.add(new Bullet(main,
+				this.getFlyer().getX() - (int) (0.5 * getFlyer().getSize()) + this.getFlyer().getSize() / 3,
+				this.getFlyer().getY() - 20, (int) this.getFlyer().getSpeedX(), 20));
+		main.add(new Bullet(main,
+				this.getFlyer().getX() - (int) (0.5 * getFlyer().getSize()) + 2 * (this.getFlyer().getSize() / 3),
+				this.getFlyer().getY() - 20, (int) this.getFlyer().getSpeedX(), 20));
 		main.getPlainGunSound().trigger();
 	}
 }
